Implement vector String.indexOf(String)
Implement the String.indexOf(String) intrinsic for Java 8 and Java 9 JCL
for z13 and z14. This is able to detect the intrinsic methods and inline
them with a vector assembly sequence to accelerate indexOf(String) search.

For Java 8 JCL, refactor Java 8 indexOf(String) APIs and introduce
JITHelper intrinsic functions. These new JITHelper APIs will be recognized
as intrinsics and inlined with the aforementioned vector sequence.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>