Software RangeCheck Read Barrier
Renaming flag and command line option S/W Evacuate RB to S/W RangeCheck
RB. Evacuate is an action that is required by RB in Concurrent
Scavenger. It's always executed in software. What has alternative
implementation in S/W vs H/W is the trigger mechanism, which is the
range check against Evacuate area boundaries (or a set of areas).

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>