Refactor code related to method fanin
- Refactor interpreter profiling methods `getNumberofCallersAndTotalWeight`
and `getOtherBucketWeight` into one method - `getFaninInfo`.
These 2 methods are usually called together, so it makes sense
to make them one method. This also reduces the number of times
a hashtable of method samples is searched.

- Move calls that access fanin info from the front-end to resolved method.
This information is only accessed from the inliner, where we always have
a resolved method. This also makes fanin info more easily accessible for
a downstream project.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>